requirements.txt
flask
flask-cors
flask-limiter
requests
python-dotenv
